Revert "#1753 Tests for getting lock name from a bean"
