+++ spring/shedlock-spring/src/test/java/net/javacrumbs/shedlock/spring/aop/MethodProxyAopTest.java
@@ -135,14 +135,6 @@ public class MethodProxyAopTest {
     }
 
     @Test
-    public void shouldGetNameFromABean() {
-        testBean.spelBean();
-        verify(lockProvider).lock(hasParams("myLockName", 30_000, 1_000));
-        verify(simpleLock).unlock();
-        assertThat(testBean.wasMethodCalled()).isTrue();
-    }
-
-    @Test
     public void shouldFailOnNotLockedMethod() {
         assertThatThrownBy(() -> testBean.finalNotLocked()).hasMessageStartingWith("The task is not locked.");
     }
